requirements.txt
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
pandas>=2.0
ultralytics==8.3.160
paddlepaddle>=2.6.0
paddleocr>=2.7.0
python-dotenv
requests
face_recognition
cmake
# dlib (Build errors on ARM)
# dlib
fastapi==0.115.0
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
onvif-zeep==0.2.12
# OPi.GPIO (Build errors on ARM)
# OPi.GPIO
tinytuya
onnx==1.17.0
onnxscript>=0.1.0
psycopg2-binary>=2.9.0
streamlit-cookies-manager==0.2.0


